Move direct driver stub server tests to a separate file
