+++ driver/src/test/java/org/neo4j/driver/internal/DirectDriverIT.java
@@ -22,43 +22,24 @@ import org.hamcrest.CoreMatchers;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
-import org.mockito.ArgumentCaptor;
 
 import java.net.URI;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
 
 import org.neo4j.driver.internal.util.EnabledOnNeo4jWith;
-import org.neo4j.driver.v1.Config;
 import org.neo4j.driver.v1.Driver;
 import org.neo4j.driver.v1.GraphDatabase;
-import org.neo4j.driver.v1.Logger;
-import org.neo4j.driver.v1.Record;
 import org.neo4j.driver.v1.Session;
 import org.neo4j.driver.v1.StatementResult;
-import org.neo4j.driver.v1.Transaction;
 import org.neo4j.driver.v1.util.DatabaseExtension;
-import org.neo4j.driver.v1.util.StubServer;
 
-import static java.util.Arrays.asList;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.core.IsEqual.equalTo;
 import static org.hamcrest.junit.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.atLeastOnce;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 import static org.neo4j.driver.internal.BoltServerAddress.LOCAL_DEFAULT;
 import static org.neo4j.driver.internal.util.Matchers.directDriverWithAddress;
 import static org.neo4j.driver.internal.util.Neo4jFeature.CONNECTOR_LISTEN_ADDRESS_CONFIGURATION;
-import static org.neo4j.driver.v1.Values.parameters;
 import static org.neo4j.driver.v1.util.Neo4jRunner.DEFAULT_AUTH_TOKEN;
-import static org.neo4j.driver.v1.util.StubServer.INSECURE_CONFIG;
 
 class DirectDriverIT
 {
@@ -130,57 +111,6 @@ class DirectDriverIT
     }
 
     @Test
-    void shouldBeAbleRunCypher() throws Exception
-    {
-        // Given
-        StubServer server = StubServer.start( "return_x.script", 9001 );
-        URI uri = URI.create( "bolt://127.0.0.1:9001" );
-        int x;
-
-        // When
-        try ( Driver driver = GraphDatabase.driver( uri, INSECURE_CONFIG ) )
-        {
-            try ( Session session = driver.session() )
-            {
-                Record record = session.run( "RETURN {x}", parameters( "x", 1 ) ).single();
-                x = record.get( 0 ).asInt();
-            }
-        }
-
-        // Then
-        assertThat( x, equalTo( 1 ) );
-
-        // Finally
-        assertThat( server.exitStatus(), equalTo( 0 ) );
-    }
-
-    @Test
-    void shouldSendMultipleBookmarks() throws Exception
-    {
-        StubServer server = StubServer.start( "multiple_bookmarks.script", 9001 );
-
-        List<String> bookmarks = Arrays.asList( "neo4j:bookmark:v1:tx5", "neo4j:bookmark:v1:tx29",
-                "neo4j:bookmark:v1:tx94", "neo4j:bookmark:v1:tx56", "neo4j:bookmark:v1:tx16",
-                "neo4j:bookmark:v1:tx68" );
-
-        try ( Driver driver = GraphDatabase.driver( "bolt://localhost:9001", INSECURE_CONFIG );
-              Session session = driver.session( bookmarks ) )
-        {
-            try ( Transaction tx = session.beginTransaction() )
-            {
-                tx.run( "CREATE (n {name:'Bob'})" );
-                tx.success();
-            }
-
-            assertEquals( "neo4j:bookmark:v1:tx95", session.lastBookmark() );
-        }
-        finally
-        {
-            assertEquals( 0, server.exitStatus() );
-        }
-    }
-
-    @Test
     void shouldConnectIPv6Uri()
     {
         // Given
@@ -192,41 +122,6 @@ class DirectDriverIT
 
             // Then
             assertThat( result.single().get( 0 ).asInt(), CoreMatchers.equalTo( 1 ) );
-        }
-    }
-
-    @Test
-    void shouldLogConnectionIdInDebugMode() throws Exception
-    {
-        StubServer server = StubServer.start( "hello_run_goodbye.script", 9001 );
-
-        Logger logger = mock( Logger.class );
-        when( logger.isDebugEnabled() ).thenReturn( true );
-
-        Config config = Config.build()
-                .withLogging( ignore -> logger )
-                .withoutEncryption().toConfig();
-
-        try ( Driver driver = GraphDatabase.driver( "bolt://localhost:9001", config );
-              Session session = driver.session() )
-        {
-            List<String> names = session.run( "MATCH (n) RETURN n.name" ).list( record -> record.get( 0 ).asString() );
-            assertEquals( asList( "Foo", "Bar" ), names );
-
-            ArgumentCaptor<String> messageCaptor = ArgumentCaptor.forClass( String.class );
-            verify( logger, atLeastOnce() ).debug( messageCaptor.capture(), any() );
-
-            Optional<String> logMessageWithConnectionId = messageCaptor.getAllValues()
-                    .stream()
-                    .filter( line -> line.contains( "bolt-123456789" ) )
-                    .findAny();
-
-            assertTrue( logMessageWithConnectionId.isPresent(),
-                    "Expected log call did not happen. All debug log calls:\n" + String.join( "\n", messageCaptor.getAllValues() ) );
-        }
-        finally
-        {
-            assertEquals( 0, server.exitStatus() );
         }
     }
 }
